refac[frontend]: split darkModeToggle component
